use DI for example; update docs
