--- SimpleDatabase/src/test/java/com/tinaciousdesign/simpledatabase/SimpleDatabaseTest.kt
@@ -1,60 +1,0 @@
-package com.tinaciousdesign.simpledatabase
-
-import org.junit.After
-import org.junit.Before
-import org.junit.Test
-
-internal class SimpleDatabaseTest {
-
-    @Before
-    fun setUp() {
-    }
-
-    @After
-    fun tearDown() {
-    }
-
-    @Test
-    fun getString() {
-    }
-
-    @Test
-    fun putString() {
-    }
-
-    @Test
-    fun putObject() {
-    }
-
-    @Test
-    fun getObject() {
-    }
-
-    @Test
-    fun getListObject() {
-    }
-
-    @Test
-    fun putListObject() {
-    }
-
-    @Test
-    fun getInt() {
-    }
-
-    @Test
-    fun putInt() {
-    }
-
-    @Test
-    fun getBoolean() {
-    }
-
-    @Test
-    fun putBoolean() {
-    }
-
-    @Test
-    fun remove() {
-    }
-}
